+++ CHANGELOG.md
@@ -7,6 +7,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 cluster API. See the changelog of the [cluster API](https://cluster-api.cyberfusion.nl/redoc#section/Changelog) for 
 detailed information.
 
+## Unreleased
+
+## Added
+
+- Apply path validation on several setters.
+- Add ability to validate multiple values, just call `->each()` after initializing the validator with a value.
+
+## Removed
+
+- Remove `positiveInteger` validation.
+
 ## [1.71.0]
 
 ### Added