+++ CHANGELOG.md
@@ -9,13 +9,13 @@ detailed information.
 
 ## [1.72.0]
 
-## Added
+### Added
 
 - Apply path validation on several setters.
 - Add ability to validate multiple values, just call `->each()` after initializing the validator with a value.
 - Pattern validation to NodeJS versions.
 
-## Removed
+### Removed
 
 - Remove `positiveInteger` validation.