+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 cluster API. See the changelog of the [cluster API](https://cluster-api.cyberfusion.nl/redoc#section/Changelog) for 
 detailed information.
 
+## [1.72.1]
+
+### Fixed
+
+- Unix user might not have a default nodejs version but validation was not allowing `null`.
+
 ## [1.72.0]
 
 ### Added